add test folder and test file
